fix(calendar): land calendar UI content missed by prior rename-only commit
The prior commit captured only the git-mv renames; the actual content
(CalendarSection.tsx, ui.ts barrel, EventEditorDrawer side-effect import,
the thin LifeOpsCalendarSection wrapper, repointed importers, ./ui export)
was left uncommitted by a working-tree race. This lands it.
